:test_tube: actually run events retry test
the test had been silently failing in the background
because the queue consumer started did not have
the di preference overriden for Config to TestConfig.

because that always happened in a separate process, it
never came to light that we weren't actually asserting
that retries had been performed in the background.